댓글 수정 삭제 ajax 적용

+++ src/main/java/com/example/boardproject/controller/BoardController.java
@@ -234,7 +234,18 @@ public class BoardController {
 
     // 게시물 삭제
     @DeleteMapping("board/{boardSeq}")
-    public String deleteBoard(Model model, @PathVariable("boardSeq") Long boardSeq) {
+    public String deleteBoard(Model model, @PathVariable("boardSeq") Long boardSeq, HttpServletRequest httpServletRequest) {
+        BoardResponseDto boardResponseDto = boardService.findByBoardSeq(boardSeq);
+        HttpSession session = httpServletRequest.getSession();
+        // 현재 로그인 중인 유저 권한
+        boolean isAdmin = (boolean) session.getAttribute("admin");
+
+        // 게시글에 대한 권한이 없으면 리스트로 돌아감
+        if(!isAdmin && !boardResponseDto.getUserSeq().equals(session.getAttribute("userSeq"))){
+            model.addAttribute("message", "삭제 권한이 없습니다.");
+            model.addAttribute("url", "/board");
+            return "/alert";
+        }
         log.info("delete board   boardSeq: {}", boardSeq);
         boardService.deleteBoard(boardSeq);
 
@@ -246,6 +257,7 @@ public class BoardController {
     // 게시물 업데이트 페이지 이동
     @GetMapping("board/edit/{boardSeq}")
     public String updateBoardPage(Model model, @PathVariable("boardSeq") Long boardSeq, HttpServletRequest httpServletRequest) {
+
         BoardResponseDto boardResponseDto = boardService.findByBoardSeq(boardSeq);
         UserDto userDto = userService.findByUserSeq(boardResponseDto.getUserSeq());
         String boardUserId = userDto.getUserId();
@@ -253,6 +265,13 @@ public class BoardController {
         HttpSession session = httpServletRequest.getSession();
         // 현재 로그인 중인 유저 권한
         boolean isAdmin = (boolean) session.getAttribute("admin");
+
+        // 게시글에 대한 권한이 없으면 리스트로 돌아감
+        if(!isAdmin && !boardResponseDto.getUserSeq().equals(session.getAttribute("userSeq"))){
+            model.addAttribute("message", "수정 권한이 없습니다.");
+            model.addAttribute("url", "/board");
+            return "/alert";
+        }
 
         if (boardResponseDto.isImageYn()) {
             String imagePath = boardService.getImagePathByBoardSeq(boardResponseDto.getBoardSeq());
@@ -274,6 +293,7 @@ public class BoardController {
                               HttpServletRequest httpServletRequest,
                               RedirectAttributes redirectAttributes,
                               BindingResult bindingResult) {
+
 
         BoardRequestDto boardRequestDto = new BoardRequestDto(boardResponseDto);
         HttpSession session = httpServletRequest.getSession();
@@ -370,14 +390,17 @@ public class BoardController {
 
 
 
-
-    @DeleteMapping("/board/{commentSeq}")
-    public void deleteComment(@PathVariable("commentSeq") Long commentSeq){
+    @ResponseBody
+    @DeleteMapping("/board/comment/{commentSeq}")
+    public Long deleteComment(@PathVariable("commentSeq") @RequestBody Long commentSeq){
         boardService.deleteComment(commentSeq);
+        return commentSeq;
     }
 
-    @PutMapping("/board/{commentSeq}")
-    public void updateComment(@PathVariable("commentSeq") Long commentSeq, BoardCommentRequestDto commentDto) {
-        boardService.updateComment(commentSeq, commentDto);
+    @ResponseBody
+    @PutMapping("/board/comment/{commentSeq}")
+    public String updateComment(@PathVariable("commentSeq") Long commentSeq, @RequestBody String commentContent) {
+        boardService.updateComment(commentSeq, commentContent);
+        return commentContent;
     }
 }
